fix credit capacity handling with market deposited collateral
when there is market deposited collateral, it should be possible for
creditCapacity to go negative in order to protect against overruns when
more collateral is withdrawn from the market than capacity provided by
the pools.

when there is no deposited collateral, then creditcapacity functions
same as before.

+++ protocol/synthetix/contracts/storage/Market.sol
@@ -75,8 +75,11 @@ library Market {
          *
          * The Market's credit capacity also has a dependency on the external market reported debt as it will respond to that debt (and hence change the credit capacity if it increases or decreases)
          *
+         * The credit capacity can go negative if all of the collateral provided by pools is exhausted, and there is market provided collateral available to consume. in this case, the debt is still being
+         * appropriately assigned, but the market has a dynamic cap based on deposited collateral types.
+         *
          */
-        uint128 creditCapacityD18;
+        int128 creditCapacityD18;
         /**
          * @dev The total balance that the market had the last time that its debt was distributed.
          *
@@ -243,12 +246,12 @@ library Market {
         Data storage self,
         uint256 creditCapacitySharesD18,
         int256 maxShareValueD18
-    ) internal view returns (uint256 contributionD18) {
+    ) internal view returns (int256 contributionD18) {
         // Determine how much the current value per share deviates from the maximum.
         uint256 deltaValuePerShareD18 = (maxShareValueD18 -
             self.poolsDebtDistribution.getValuePerShare()).toUint();
 
-        return deltaValuePerShareD18.mulDecimal(creditCapacitySharesD18);
+        return deltaValuePerShareD18.mulDecimal(creditCapacitySharesD18).toInt();
     }
 
     /**
@@ -256,7 +259,7 @@ library Market {
      *
      */
     function isCapacityLocked(Data storage self) internal view returns (bool) {
-        return self.creditCapacityD18 < getLockedCreditCapacity(self);
+        return self.creditCapacityD18 < getLockedCreditCapacity(self).toInt();
     }
 
     /**
